+++ tests/testthat/test-run_shiny.R
@@ -1,5 +1,5 @@
 
 test_that("Shiny app is generated", {
   load(system.file('extdata/tiny_rse.Rdata', package = 'geyser'))
-  expect_is(geyser(tiny_rse,'Test RSE'), "shiny.appobj")
+  expect_type(geyser(tiny_rse,'Test RSE'), "list")
 })